Update README.md
Our team received a request to improve the recommendation system on our site, Amantes da Música. This site has a music recommender that uses only the music genre to make recommendations, but they are not pleasing the visitors.

Our challenge, therefore, is to improve the recommendation system, using clustering and dimensionality reduction techniques. For this, we will use some PySpark modules, such as SQL, to manipulate the data, and Machine Learning, which will allow the use of various tools aimed at clustering solutions.